[genmetadata] Execute find cd'ed to the subvolume
This avoids having the subvolume's name spelt out in the stat log file
again and again.

As it would be difficult to impossible (depending on the situation)
to give a relative path from cd'ed to the subvolume to the log file
outside, the log file is opened by the shell script and passed as a
file descriptor number in the environment, instead.