[genmetadata] Provide optional infix via --infix=INFIX
This is meant to create multiple sets of meta-data files
for the same subvolume. (e.g., it may not be read-only
and changes should get reflected in new meta-data; or
to prepare comparing between meta-data copied over
from the source host and meta-data taken
from the target host.)

View diff with -w (ignore white-space).