Add cvnsnapper-replicate-statusmetadata (missing main functionality)
This will compare the collected meta-data a bit like "snapper status",
but without needing access to the full subvolume at hand. Thus,
it can be used to give status information between original subvolume
on the source host, and replicated subvolume on the destination host.

It's just reading in the data, for now. It'll still need to sort
both sides, then compare them and output the differences.

The script is losely based on the inline Perl script in sort-stat
in cvnsnapper-replicate-genmetadata.