Add cvnsnapper-import-name2infoxml
This is for converting (some forms of) a manually created snapshot's name
to a meta-data info.xml for constructing/faking a .snapshots hierarchy
snapper will be able to work with.

Currently, the input name must be in the form of:

  YYYYMMDD-hhmmss-${DESCRIPTION}

.., with the date-time being in localtime (whereas snapper info.xml date
seems to be in UTC always). The description may have underscores
instead of spaces.

The name will be split into date-time and description; then the
date-time demangled to a form that date --date=... can parse,
and converted to "YYYY-MM-DD hh:mm:ss" UTC.

A number will count up from 1, and all the information is then
interpolated into a snapper info.xml template by an inline Perl script.
(It may do some escaping to XML entities in the process.)
This is output to stdout.

An example run:

  fabian@falcon:~/opt/cvnsnapper-toolbox$ ./cvnsnapper-import-name2infoxml \
    20150311-172041-after_copying_from_hdgen2
  <?xml version="1.0"?>
  <snapshot>
    <type>single</type>
    <num>1</num>
    <date>2015-03-11 16:20:41</date>
    <uid>0</uid>
    <description>after copying from hdgen2</description>
  </snapshot>

(Note the difference between 172041 local time and 16:20:41 UTC.)