[plain-statusmetadata] Bugfix: Don't include previous record number in list of stat records
The intention was a comma operator as in C,
which would evaluate the left side and throw it away...

But here in Perl a statement separator is better, as (only)
the last statement will implicitly be used as return value
of the map block if there is no explicit return. In other words,
the $record_num++ will be ignored for block result and not included
anymore in the overall map result.