[plain-statusmetadata] Add option --ignore-profile={same,snapshots,replicated}
Also add a description for each of those possible option arguments,
but they are not user-accessible at the moment.

--ignore-profile and --add-ignore-profile add to the previous list
of ignore fields, while --set-ignore-profile overrides the previous list.

The idea is to give a tool into the hand of the user to easily/quickly
add/change ignore fields, which is essential for getting a clean/useful
status comparison. It should also help the user avoid having to think hard
about which fields in particular to ignore. If needed, a detailed list
of fields to ignore can always be given via --(/add-/set-)ignore-fields=,
though, as before.