[plain-backingstorage] Be quieter by default: Insert a new verbose level -v; move previous level -v to -v -v
It was too long, too many information thrown together in one place.
So, try to output at basic/default verbose level only what is supposed
to be critically important; what is supposed to give access to the
most-of-the-time "hidden" information that could help in
descision making, or just give a warm feeling when used to verify
that, in fact, you're using the right partitions with dd, in the right
direction.