[plain-backingstorage] ZFS support: Try to avoid echo for passing on arbitrary data
echo could see a leading dash as an option. So, use "<<<" instead,
possibly combined with cat to preserve the left-to-right "flow"
of a line. (Replace the echo that is assured to begin with an
alphanumeric word, too, for good measure(?).)

Also transform use of sed a bit.
* Use other delimeters when handling slash '/', for easier readability.
* Use -e wherever an inline script is given, to be safe(?).
* Merge use of grep with following sed where that was possible.

As I don't have ZFS myself, I hope everything still works!