[plain-backingstorage] Add mode show-holders, which is reverse show
Up until now, there was only the show mode, which provides insight
into what is the backing storage of a given file-system object
or block device.

With the new mode show-holders, it can be seen what a given
backing storage is currently being used by. That is roughly
the reverse of mode show, as not the extracted parent information
is followed, but the holders information from sysfs.