replicate-send: For --receive, don't skip timeline snapshots
It's meant to give automatic incremental replication, one day.
Would need to work out / confirm-by-user the deletion of (some)
timeline snapshots at the source, and to replicate it at the
destination, though.

For manual cold storage, ignoring timeline would be good as
they's mostly not permanent, and a manual snapshot would have been
made shortly before. For automatic "live" data replication,
this doesn't hold; let's try to replicate as much as we can!