replicate-send: Use liblog.sh; only dump snapshot meta-data on -v
While this might be interesting while waiting for the send operation
to complete, it can also be overwhelming to see all the meta-data
all the time. So only show it with -v (--verbose).

As this might mean that there is no detailed snapshot number/info
anymore, include the current snapshot number in the pre-send
status output.

That is:

    -		echo -n "Full send: "
    +		echo -n "Full send of $SNAPSHOT_NUMBER: "

    -		echo -n "Incremental send with parent $LAST_SNAPSHOT_NUMBER: "
    +		echo -n "Incremental send from $LAST_SNAPSHOT_NUMBER to $SNAPSHOT_NUMBER: "